+++ apps/web/components/user-profile/modules/add-done.tsx
@@ -1,44 +1,20 @@
 import { Pages } from 'types'
 import { SettingsSection } from '@/ui/settings/lists/base'
-import { dashed } from '@/utils/array'
-import { Row } from '@/ui/settings/lists/rows'
 import { Button } from '@/ui/buttons'
 import { SetState } from '@modtree/types'
 import { SettingsSearchBox } from '@/ui/search'
-import { ModuleCondensed } from '@modtree/entity'
 import { useAppDispatch, useAppSelector } from '@/store/redux'
-import { removeFromBuildList } from '@/store/search'
 import { updateModulesDone } from '@/store/user'
-
-function SelectedModules(props: { modules: ModuleCondensed[] }) {
-  const dispatch = useAppDispatch()
-  return (
-    <>
-      {props.modules.length !== 0 && (
-        <div className="ui-rectangle flex flex-col overflow-hidden">
-          {props.modules.map((module, index) => (
-            <Row.Module
-              key={dashed(module.moduleCode, index)}
-              deletable
-              onDelete={() => dispatch(removeFromBuildList(module))}
-            >
-              <span className="font-semibold">{module.moduleCode}</span>
-              <span className="mx-1">/</span>
-              {module.title}
-            </Row.Module>
-          ))}
-        </div>
-      )}
-    </>
-  )
-}
+import { SelectedModules } from './selected-modules'
+import { updateCachedModulesCondensed } from '@/utils/backend'
 
 export function AddDone(props: { setPage: SetState<Pages['Modules']> }) {
   const dispatch = useAppDispatch()
   const buildList = useAppSelector((state) => state.search.buildList)
   function confirm() {
-    console.log('confirm', buildList)
-    dispatch(updateModulesDone(buildList.map((m) => m.moduleCode)))
+    const codes = buildList.map((m) => m.moduleCode)
+    dispatch(updateModulesDone(codes))
+    updateCachedModulesCondensed(dispatch, codes)
     props.setPage('main')
   }
   return (
@@ -62,7 +38,7 @@ export function AddDone(props: { setPage: SetState<Pages['Modules']> }) {
       </SettingsSection>
       <div className="flex flex-row-reverse">
         <Button color="green" onClick={() => confirm()}>
-          Add to modules done
+          Save changes
         </Button>
       </div>
     </div>
